Move TimerInternals to runners/core-java
